backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
passlib[bcrypt]==1.7.4
python-jose